Move getPitchMatrix tests to thier relevant folder
