+++ src/harpstrata/Pitch/index.test.ts
@@ -1,11 +1,6 @@
-import { MAJOR_DIATONIC_APPARATUS } from '../Apparatus/constants'
-
-import { EXAMPLE_PITCH_MATRICES } from './testResources'
-
 import { PitchIds } from './types'
 import type { Pitch } from './types'
-import { getPitchMatrix, getActivePitchIds, getPitch } from './index'
-import { C, F } from './constants'
+import { getActivePitchIds, getPitch } from './index'
 
 
 test('getActivePitchIds function returns an array of the available pitches', () => {
@@ -20,25 +15,4 @@ test('getPitch function can return a first pozition', () => {
   const actualPitch = getPitch(C_PITCH.id)
 
   expect(actualPitch).toStrictEqual(C_PITCH)
-})
-
-test('getPitchMatrix function maps a simple 2d array of 0\'s to the input key pitch of C', () => {
-  const expectedArray = [[ C, ], [ C, ]]
-  const actualArray = getPitchMatrix([[ 0, ], [ 0, ]], C.id)
-
-  expect(actualArray).toStrictEqual(expectedArray)
-})
-
-test('getPitchMatrix maps a major diatonic halfstepmatrix in to a major diatonic pitchMatrix for a C harmonica', () => {
-  const { MAJOR_DIATONIC_C_HARMONICA } = EXAMPLE_PITCH_MATRICES
-  const actualArray = getPitchMatrix(MAJOR_DIATONIC_APPARATUS.halfstepIndexMatrix, C.id)
-
-  expect(actualArray).toStrictEqual(MAJOR_DIATONIC_C_HARMONICA)
-})
-
-test('getPitchMatrix maps a major diatonic halfstepmatrix in to a major diatonic pitchMatrix for a F harmonica', () => {
-  const { MAJOR_DIATONIC_F_HARMONICA } = EXAMPLE_PITCH_MATRICES
-  const actualArray = getPitchMatrix(MAJOR_DIATONIC_APPARATUS.halfstepIndexMatrix, F.id)
-
-  expect(actualArray).toStrictEqual(MAJOR_DIATONIC_F_HARMONICA)
 })
